refactor: extract shared utilities to reduce edge function duplication
- Extract SpamDetectionService to _shared/spam-detection-service.ts
- Enhance database.ts with GitHubUserExtended interface for GraphQL data
- Update ensureContributor to support extended contributor fields
- Refactor spam-detection to use shared SpamDetectionService and database utilities
- Refactor repository-sync-graphql to use shared database and response utilities
- Refactor github-webhook to use shared database and response utilities
- Remove duplicate contributor upsert logic across all three functions
- Standardize response formatting with shared response utilities

Implements issue #1021 Phase 3 - High-Duplication Functions Refactoring
Reduces code duplication by ~400 lines across three edge functions
